+++ src/main/java/com/codeborne/selenide/webdriver/FirefoxDriverFactory.java
@@ -2,6 +2,7 @@ package com.codeborne.selenide.webdriver;
 
 import com.codeborne.selenide.Browser;
 import com.codeborne.selenide.Config;
+import org.apache.commons.io.IOUtils;
 import org.openqa.selenium.Proxy;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
@@ -10,7 +11,13 @@ import org.openqa.selenium.firefox.FirefoxProfile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.Optional;
+
+import static java.lang.Boolean.parseBoolean;
+import static java.lang.Integer.parseInt;
+import static java.nio.charset.StandardCharsets.UTF_8;
 
 class FirefoxDriverFactory extends AbstractDriverFactory {
   private static final Logger log = LoggerFactory.getLogger(FirefoxDriverFactory.class);
@@ -49,31 +56,60 @@ class FirefoxDriverFactory extends AbstractDriverFactory {
     firefoxOptions.addPreference("network.proxy.allow_hijacking_localhost", true);
 
     firefoxOptions.merge(createCommonCapabilities(config, proxy));
-    firefoxOptions = transferFirefoxProfileFromSystemProperties(firefoxOptions);
+
+    FirefoxProfile profile = Optional.ofNullable(firefoxOptions.getProfile()).orElseGet(FirefoxProfile::new);
+    setupDownloadsFolder(config, profile);
+    transferFirefoxProfileFromSystemProperties(profile);
+    firefoxOptions.setProfile(profile);
 
     return firefoxOptions;
   }
 
-  private FirefoxOptions transferFirefoxProfileFromSystemProperties(FirefoxOptions currentFirefoxOptions) {
+  private void setupDownloadsFolder(Config config, FirefoxProfile profile) {
+    if (profile.getStringPreference("browser.download.dir", "").isEmpty()) {
+      profile.setPreference("browser.download.dir", new File(config.downloadsFolder()).getAbsolutePath());
+      profile.setPreference("browser.helperApps.neverAsk.saveToDisk", popularContentTypes());
+      profile.setPreference("pdfjs.disabled", true);  // disable the built-in viewer
+      profile.setPreference("browser.download.folderList", 2); // 0=Desktop, 1=Downloads, 2="reuse last location"
+    }
+  }
+
+  String popularContentTypes() {
+    try {
+      return String.join(";", IOUtils.readLines(getClass().getResourceAsStream("/content-types.properties"), UTF_8));
+    }
+    catch (IOException e) {
+      return "text/plain;text/csv;application/zip;application/pdf;application/octet-stream;" +
+        "application/msword;application/vnd.ms-excel;text/css;text/html";
+    }
+  }
+
+  private void transferFirefoxProfileFromSystemProperties(FirefoxProfile profile) {
     String prefix = "firefoxprofile.";
-    FirefoxProfile profile = Optional.ofNullable(currentFirefoxOptions.getProfile())
-      .orElseGet(FirefoxProfile::new);
+
     for (String key : System.getProperties().stringPropertyNames()) {
       if (key.startsWith(prefix)) {
         String capability = key.substring(prefix.length());
         String value = System.getProperties().getProperty(key);
         log.debug("Use {}={}", key, value);
-        if (value.equals("true") || value.equals("false")) {
-          profile.setPreference(capability, Boolean.valueOf(value));
+        if (isBoolean(value)) {
+          profile.setPreference(capability, parseBoolean(value));
         }
-        else if (value.matches("^-?\\d+$")) { //if integer
-          profile.setPreference(capability, Integer.parseInt(value));
+        else if (isInteger(value)) {
+          profile.setPreference(capability, parseInt(value));
         }
         else {
           profile.setPreference(capability, value);
         }
       }
     }
-    return currentFirefoxOptions.setProfile(profile);
+  }
+
+  private boolean isBoolean(String value) {
+    return "true".equalsIgnoreCase(value) || "false".equalsIgnoreCase(value);
+  }
+
+  private boolean isInteger(String value) {
+    return value.matches("^-?\\d+$");
   }
 }